% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountReg.R
\docType{package}
\name{CountReg}
\alias{CountReg}
\alias{CountReg-package}
\title{Regression Models for Dispersed Count Data}
\description{
The CountReg package contains functions to modelling
    count data via likelihood-based approach with a S4
    implementation. The models considered are Generalized-Poisson,
    COM-Poisson and Gamma-Count and methods for the models also are
    provided.
}

