% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_plot.R
\name{dat_plot}
\alias{dat_plot}
\title{Plotting a Scatter plot Matrix of Data}
\usage{
dat_plot(data)
}
\arguments{
\item{data}{A dataframe object}
}
\description{
\code{dat_plot} is used to summarize/model data through the use of a scatter plot
matrix of a data set's variables (i.e. columns).
}
\details{
\code{dat_plot} require the installation of \code{\link{GGally}} and \code{\link{ggplot2}}. The
matrix is formated to plot/compute correlations and data summarizations based on data
structure.

NOTE: It's not recommended to use this function past 15 variables, due to complexity.
Try sub-setting data set or utilize \code{\link{ggplot2}}.
}
\examples{
## Example using bioChemists dataset from pscl (Campbell & Mahon, 1974)
utils::data(bioChemists, package = "pscl")
dat_plot(bioChemists)

## Example using mtcars data set (Motor Trend, 1974)
utils::data(mtcars)
dat_plot(cars)

}
\references{
Long, J. S. (1997).
Regression Models for Categorical and Limited Dependent Variables.
Sage Publications.
}
\author{
Implementation of \code{dat_plot} was authored by Jeremy Artiga, with aid
from William Cipolli at Colgate University.
}
