% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammacount.R
\docType{package}
\name{gammacount}
\alias{gammacount}
\alias{gammacount-package}
\title{Regression Models for Dispersed Count Data}
\description{
The gammacount is a package that contains functions to
    modelling count data via likelihood-based approach with a S4
    implementation. The Gamma-count regression models are considered
    in package and methods functions for this models also are
    provided.
}
