% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_negb.R
\name{glm_negb}
\alias{glm_negb}
\title{Generalized Linear Model: Negative Binomial Regression}
\usage{
glm_negb(data, formula, offsetparm = NULL)
}
\arguments{
\item{data}{A dataframe object that contains vectors spesified by the symbolic
representation of \code{formula}}

\item{formula}{A symbolic represenation of the linear predictor used to fit the
model}

\item{offsetparm}{String representation of a dataframe column which applies an offset
to the linear predictor}
}
\value{
\code{glm_negb} returns an S3 object of class \code{"glm_negb"}, it is not inherited from
classes \code{"lm"} or \code{"glm"}. Functions specified for these packages will not work
with a \code{"glm_negb"} object.

The function \code{\link{interpret}} can be used to obtain and print a summary of
results.

The \code{"$"} syntax can be used to extract various useful features/information
of the output values from the initial model fit. Many of the features included in
the \code{"glm_pois"} object are layered in lists, any computations utlilizing these
features should be unlisted first using \code{\link{unlist}}.

An object of class \code{"glm_negb"} contains at least the following components:

\item{coefficients}{List containing estimated coefficients and their standard errors}
\item{residuals}{Working residuals computed in the final iteration of the IRWLS fit}
\item{summary}{List containing statistical inference and features of model fit}
\item{fitted.values}{Fitted mean values (i.e., \eqn{\lambda}) obtained by exponentiation of the linear predictor}
\item{df.residual}{Degrees of freedom for residuals}
\item{call}{The matched call}
\item{terms}{The \code{\link{terms}} object used}
\item{model}{List containing the response and design matrices}
\item{theta}{\eqn{\theta} approximated as dispersion factor}
}
\description{
\code{glm_negb} is used to fit the negative binomial generalized linear model, specified
by \code{formula}, a symbolic representation of the linear predictor.
}
\details{
A negative binomial generalized linear model is a statistical model that fits count-based
data by utilization of an IRWLS algorithm that makes proper adjustments to \eqn{\beta}
until convergence.

A dispersion parameter \eqn{\theta} is also approximated within the algorithmic fit as
a dispersion control. It's typically fitted towards handling over-dispersion, but may
handle some cases of under-dispersion.

NOTE: THE \code{offsetparm}'s offset is logged before being fit, be cautious when entering offset
data
}
\examples{
## Example using mtcars data set (Motor Trend, 1974)
utils::data(mtcars)

mod <- glm_negb(data=mtcars, hp~factor(gear)+factor(carb)+disp)

## Using an offset
mod <- glm_negb(data=mtcars, hp~factor(gear)+factor(carb)+disp, offsetparm="wt")

##Extracting features/information
interpret(mod)
betas <- mod$coefficients$betas
std.err <- mod$coefficients$std.error
disp <- mod$dispersion

}
\references{
Motor Trend Magazine (1974).
Motor Trend Road Tests.
}
\author{
Implementation of \code{glm_negb} was authored by Jeremy Artiga, with aid
from William Cipolli at Colgate University.
}
