% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_negb_zero.R
\name{glm_negb_zero}
\alias{glm_negb_zero}
\title{Generalized Linear Model: Zero-Inflated Negative Binomial Regression}
\usage{
glm_negb_zero(data, formula.negb, formula.log, offsetparm = NULL)
}
\arguments{
\item{data}{A dataframe object that contains vectors specified by the symbolic
representation of \code{formula}}

\item{formula.negb}{A symbolic representation of the linear predictor used to fit the
poisson model}

\item{formula.log}{A symbolic representation of the linear predictor used to fit the
logistic model}

\item{offsetparm}{String representation of a dataframe column which applies an offset
to the linear predictor on \strong{BOTH} models.}
}
\value{
\code{glm_negb_zero} returns an S3 object of class \code{"glm_negb_zero"}, it is not inherited from
classes \code{"lm"} or \code{"glm"}. Functions specified for these packages will not work
with a \code{"glm_negb_zero"} object.

The function \code{\link{interpret}} can be used to obtain and print a summary of
results.

The \code{"$"} syntax can be used to extract various useful features/information
of the output values from the initial model fit. Many of the features included in
the \code{"glm_negb_zero"} object are layered in lists, any computations utlilizing these
features should be unlisted first using \code{\link{unlist}}.

An object of class \code{"glm_negb_zero"} contains at least the following components:

\item{coefficients}{List containing estimated coefficients for poisson/logistic models
and their standard errors}
\item{residuals}{Working residuals for the poisson model computed in the final iteration of the IRWLS fit}
\item{summary}{List containing statistical inference and features of poisson/logistic model fit}
\item{fitted.values}{Fitted mean values (i.e., \eqn{\lambda}) obtained by exponentiation of the poisson linear predictor}
\item{pred.zero}{Fitted values (i.e., \eqn{\alpha}) obtained by exponentiation of the logistic linear predictor}
\item{df.residual}{Degrees of freedom for residuals}
\item{call}{The matched call}
\item{terms}{The \code{\link{terms}} object used}
\item{model}{List containing the response and design matrices}
\item{theta}{\eqn{\theta} approximated as dispersion factor}
}
\description{
\code{glm_pois_zero} is used to fit the zero-inflated negative binomial generalized linear model, specified
by \code{formula}, a symbolic representation of the linear predictor.
}
\details{
A zero-inflated negative binomia generalized linear model is a statistical model that fits count-based
data by utilization of an IRWLS algorithm that makes proper adjustments to \eqn{\beta} of the
poisson model and \eqn{\alpha} of the logistic model until convergence.

A dispersion parameter \eqn{\theta} is also approximated within the algorithmic fit as
a dispersion control. It's typically fitted towards handling over-dispersion, but may
handle some cases of under-dispersion.

The logistic model is spesified by:
\deqn{log(\frac{P(y_i=0)}{1-P(y_i=0)})=\alpha_0+\alpha_1 x_1 + \cdots + \alpha_{n-1} x_{n-1}}
Which models the probablity of observing structural zeros.

NOTE: THE \code{offsetparm}'s offset is logged before being fit, be cautious when entering offset
data
}
\examples{
## Example using NMES1988 dataset from AER (Deb & Trivedi, 1997)
utils::data(NMES1988, package = "AER")

mod <- glm_negb_zero(data=NMES1988, visits~factor(health), visits~factor(health))

## With offset
mod <- glm_negb_zero(data=NMES1988, visits~factor(health), visits~factor(health), offsetparm="age")

##Extracting features/information
interpret(mod)
betas <- mod$coefficients$count
std.err <- mod$coefficients$std.error.count
zeros <- mod$pred.zero

}
\references{
Deb, P., & Trivedi, P. K. (1997).
Demand for medical care by the elderly: A finite mixture approach.
Journal of Applied Econometrics, 12(3), 313--336.
}
\author{
Implementation of \code{glm_negb_zero} was authored by Jeremy Artiga, with aid
from William Cipolli at Colgate University.
}
