% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_pois_GP2.R
\name{glm_pois_GP2}
\alias{glm_pois_GP2}
\title{Generalized Linear Model: Generalized Poisson (GP2)}
\usage{
glm_pois_GP2(data, formula, offsetparm = NULL)
}
\arguments{
\item{data}{A dataframe object that contains vectors spesified by the symbolic
representation of \code{formula}}

\item{formula}{A symbolic represenation of the linear predictor used to fit the
model}

\item{offsetparm}{String representation of a dataframe column which applies an offset
to the linear predictor}
}
\value{
\code{glm_pois_GP2} returns an S3 object of class \code{"glm_pois_GP2"}, it is not inherited from
classes \code{"lm"} or \code{"glm"}. Functions specified for these packages will not work
with a \code{"glm_pois_GP2"} object.

The function \code{\link{interpret}} can be used to obtain and print a summary of
results.

The \code{"$"} syntax can be used to extract various useful features/information
of the output values from the initial model fit. Many of the features included in
the \code{"glm_pois_GP2"} object are layered in lists, any computations utlilizing these
features should be unlisted first using \code{\link{unlist}}.

An object of class \code{"glm_pois_GP2"} contains at least the following components:

\item{coefficients}{List containing estimated coefficients and their standard errors}
\item{residuals}{Working residuals computed in the final iteration of the IRWLS fit}
\item{summary}{List containing statistical inference and features of model fit}
\item{fitted.values}{Fitted mean values (i.e., \eqn{\lambda}) obtained by exponentiation of the linear predictor}
\item{df.residual}{Degrees of freedom for residuals}
\item{call}{The matched call}
\item{terms}{The \code{\link{terms}} object used}
\item{model}{List containing the response and design matrices}
\item{alpha}{\eqn{\alpha} approximated as dispersion factor}
}
\description{
\code{glm_pois_GP2} is used to fit the Generalized poisson (GP2) generalized linear model, specified
by \code{formula}, a symbolic representation of the linear predictor.
}
\details{
A generalized poisson (GP2) generalized linear model is a statistical model that fits count-based
data by utilization of an IRWLS algorithm that makes proper adjustments to \eqn{\beta}
until convergence.

A dispersion parameter \eqn{\alpha} is also approximated within the algorithmic fit as
a dispersion control. It's typically fitted to handle all dispersion types.

The following probability model is utilized through \code{\link{VGAM}}'s \code{\link{dgenpois2}}
function for the approximation of \eqn{\alpha}.
\deqn{P_y(Y_i=y_i)=\frac{\lambda(\lambda+\alpha y)^{y-1}}{y_i!}\mathrm{exp}^{-(\lambda+\alpha y)}}

NOTE: THE \code{offsetparm}'s offset is logged before being fit, be cautious when entering offset
data
}
\examples{
## Example using warpbreaks data set (British Textile Research data, early loom breakage experiments)
#' utils::data(warpbreaks)

mod <- glm_pois_GP2(data=warpbreaks, breaks~factor(wool)+factor(tension))

## Using an "example" offset
off <- rpois(length(warp), 5)
warp <- cbind(warpbreaks, off)
mod <- glm_pois_GP2(data=warp, breaks~factor(wool)+factor(tension), offsetparm="off")

##Extracting features/information
interpret(mod)
betas <- mod$coefficients$betas
std.err <- mod$coefficients$std.error
disp <- mod$dispersion

}
\references{
Bailey, N. T. J. (1953).
Statistical Methods in Biology.
Cambridge University Press.
}
\author{
Implementation of \code{glm_pois_GP2} was authored by Jeremy Artiga, with aid
from William Cipolli at Colgate University.
}
