% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare.R
\name{mod_compare}
\alias{mod_compare}
\title{Comparison of Model Fit}
\usage{
mod_compare(mod1, mod2, print.results = T)
}
\arguments{
\item{mod1}{First Model of classes built by \code{\link{countmods}}}

\item{mod2}{Second Model of classes built by \code{\link{countmods}}}

\item{print.results}{T or F depending on user prefrence for result display. It
is recommended to keep this parameter as T.}
}
\value{
List object that contains AICs, BICs, Likelihoods, and McFadden's \eqn{R^2}
for both models
}
\description{
\code{mod_compare} is used to compare models from \code{\link{countmods}} for
goodness-of-fit using various model fit metrics
}
\details{
\code{mod_compare} utilizes AIC, BIC, Likelihood, and McFadden's \eqn{R^2} as
metric of goodness-of-fit model comparisons. This may also be used as a generic
function for computation of these statistics by passing the same model to \code{mod1}
and \code{mod2}.

A likelihood ratio test is also carried out. Twice the difference of likelihoods
is compared to the Chi-squared distribution.

The user may also have the option of returning just the statistics related to
standard model comparison (AIC, BIC, Likelihood, McFadden) without displaying results
using \code{print.results}. Evaluating \code{print.results} as FALSE will compute the
likelihood ratio test, but will not display results.
}
\examples{
## Example using NMES1988 dataset from AER (Deb & Trivedi, 1997)
utils::data(NMES1988, package = "AER")
mod1 <- glm_pois(data=NMES1988, visits~factor(chronic)+factor(health))
mod2 <- glm_negb(data=NMES1988, visits~factor(chronic)+factor(health))

comp.stats <- mod_compare(mod1, mod2)

## Or as a generic function for statistics
gen.stats <- mod_compare(mod1, mod1, print.results=F)

}
\references{
Deb, P., & Trivedi, P. K. (1997).
Demand for medical care by the elderly: A finite mixture approach.
Journal of Applied Econometrics, 12(3), 313--336.
}
\author{
Implementation of \code{mod_compare} was authored by Jeremy Artiga, with aid
from William Cipolli at Colgate University.
}
