% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgnzp.R
\name{rgnzp}
\alias{rgnzp}
\title{Density, distribution function, quantile function and random
    generation for the Generalized Poisson}
\usage{
rgnzp(n, lambda, alpha = 0)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{lambda}{Vector that represents the mean (expected value) of the
distribution (\code{lambda > 0}).}

\item{alpha}{Vector that represents the dispersion parameter of the
distribution. Values of \code{alpha > 0} implies in
overdispersion while \code{alpha < 0} implies in
subdispersion. When \code{alpha = 0}, the Generalized Poisson
reduces to the Poisson distribution.}
}
\value{
\code{rgnzp} returns a vector of random generated numbers.
}
\description{
Formula and etc.
}
\details{
The function to generate random numbers is implemented in
    C++.
}
\examples{

rgnzp(n = 10, lambda = 5, alpha = 0)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
