% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinf_test.R
\name{zeroinf_test}
\alias{zeroinf_test}
\title{Test for Zero-Inflation of Count Regression Models}
\usage{
zeroinf_test(mod)
}
\arguments{
\item{mod}{Models of classes built by \code{\link{countmods}}}
}
\description{
\code{zeroinf_test} is used to test zero inflation within the \code{\link{countmods}} package.
}
\details{
\code{zeroinf_test} tests models in \code{\link{countmods}} through the simulation of the
expected response and counting the number of zeros found in the data. A distribution is
built through the simulations of the zero counts and the emperical zero count is plotted
on top of the distribution.
A right tailed emperical test is used to gauge the need for a zero-inflated model.

\code{zeroinf_test} provides the interpretation automatically.
}
\examples{
## Example using biochemists dataset from pscl (Campbell & Mahon, 1974)
utils::data(bioChemists, package = "pscl")
mod <- glm_pois(data=df, kid5~phd)

zeroinf_test(mod)

}
\references{
Long, J. S. (1997).
Regression Models for Categorical and Limited Dependent Variables.
Sage Publications.
}
\author{
Implementation of \code{zeroinf_test} was authored by Jeremy Artiga, with aid
from William Cipolli at Colgate University.
}
